+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Restrict Run.hash to auto-generated values only (alberttorosyan)
 
+### Fixes:
+
+- Fix tensorboard convert while converting tensor (sharathmk99)
+
 ## 3.12.0 Jul 22, 2022
 
 ### Enhancements: